separate run and print message from main
