+++ menu.rb
@@ -1,0 +1,15 @@
+def print_hello
+  puts 'Welcome to School Library App!'
+end
+
+def print_message
+  puts "\n"
+  puts "Please choose an option by enterin a number:
+  [1] - List all books
+  [2] - List all people
+  [3] - Create a person
+  [4] - Create a book
+  [5] - Create a rental
+  [6] - List all rentals for a given person id
+  [7] - Exit"
+end
